+++ src/eca/messenger.clj
@@ -9,6 +9,7 @@
   (chat-cleared [this params])
   (chat-status-changed [this params])
   (chat-deleted [this params])
+  (chat-opened [this params])
   (rewrite-content-received [this data])
   (tool-server-updated [this params])
   (config-updated [this params])